credential_manager/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
google-cloud-secret-manager==2.16.0
google-auth==2.26.0
google-cloud-iam-credentials==0.6.0
pydantic==1.10.9
requests==2.31.0
fastapi
uvicorn[standard]
google-cloud-secret-manager
google-auth
pydantic
google-api-python-client
